fix(lint): clear golangci-lint findings surfaced after the merge
The merge brought main's newer linter configuration onto this branch, which
flags a set of issues the branch had not been checked against — both in new
code and in a few pre-existing spots now inside the changed set.

Substantive: the Arcade status switch and the dashboard's action-status
switch now name every lifecycle value instead of leaning on default, so a
new status added upstream fails the build rather than silently landing in
the "other" bucket. Two shadowed err bindings in the storage repo are
renamed to say which operation they came from.

Mechanical: gofmt on network_endpoints.go, an unnecessary conversion, an
unused nolint directive, an embedded-field selector, three testifylint
argument orders, and a shadowed txID in a services test.

+++ pkg/wallet/fuelkeeper/keeper_test.go
@@ -45,7 +45,7 @@ func (f *fakeWallet) ListOutputs(_ context.Context, args sdk.ListOutputsArgs, _ 
 
 func (f *fakeWallet) FanOutFuel(_ context.Context, shape wdk.ShapedChange, _ string) (*sdk.CreateActionResult, error) {
 	f.mu.Lock()
-	if string(shape.Basket) == "reserve" && f.maxChunkCount > 0 && uint64(shape.Count) > f.maxChunkCount {
+	if string(shape.Basket) == "reserve" && f.maxChunkCount > 0 && shape.Count > f.maxChunkCount {
 		f.mu.Unlock()
 		return nil, errors.New("funding failed: not enough funds")
 	}
